move assembly related functions to separate file
